more bugs fixed related to jpa migration

+++ interfaces/src/main/java/ar/edu/itba/paw/persistence/ArtistDao.java
@@ -9,7 +9,7 @@ public interface ArtistDao extends CrudDao<Artist> {
     List<Artist> findBySongId(long id);
     List<Artist> findByNameContaining(String sub);
 
-    boolean updateRating(long artistId, float newRating, int newRatingAmount);
+    boolean updateRating(long artistId, Double newRating, int newRatingAmount);
     boolean hasUserReviewed(long userId, long artistId);
 
     boolean deleteReviewsFromArtist(long artistId);
